Expose `peel_casts` method as an util method inside `rustc_lint`
